Added TemplatePath and moved TemplateNode to separat src file
