+++ src/renderer.js
@@ -1,9 +1,7 @@
-import {select,matcher} from "d3-selection";
-import {FieldParser} from "./field-parser";
+import { select } from "d3-selection";
 
 // Globals
-var namedRenderFilters = {};
-export var fieldParser = new FieldParser();
+export var namedRenderFilters = {};
 
 // Main function
 export function renderFilter(name, filterFunc) {
@@ -31,39 +29,27 @@ function renderFilterPrivate(name, filterFunc, isTweenFilter) {
 }
 
 // Renderer - Renders data on element
-function Renderer(fieldSelectorAndFilters, elementSelector) {
-
-	// Parse field selector and (optional) filters
-	var parseResult = fieldParser.parse(fieldSelectorAndFilters);
-	if(parseResult.value === undefined) {
-		throw new SyntaxError("Failed to parse field selector and/or filter <" + fieldSelectorAndFilters + "> @ " + parseResult.index + ": " + parseResult.errorCode);
-	} else if(parseResult.index !== fieldSelectorAndFilters.length) {
-		throw new SyntaxError("Failed to parse field selector and/or filter <" + fieldSelectorAndFilters + "> @ " + parseResult.index + ": EXTRA_CHARACTERS");
-	}
-
-	// Set instance variables
-	this.dataFunction = createDataFunction(parseResult.value);
-	this.elementSelector = elementSelector;
+function Renderer(templatePath) {
+	this.templatePath = templatePath;
 }
 
 Renderer.prototype.render = function(/* templateElement */) {
 	// Intentionally left empty
 };
 
-// Answer the data function
-Renderer.prototype.getDataFunction = function() {
-	return this.dataFunction;
+// Answer the templatePath of the receiver
+Renderer.prototype.getTemplatePath = function() {
+	return this.templatePath;
 };
 
-// Answer the element which should be rendered (indicated by the receivers elementSelector)
-Renderer.prototype.getElement = function(templateElement) {
+// Answer the element referred to by the receivers templatePath
+Renderer.prototype.getElementIn = function(rootElement) {
+	return this.getTemplatePath().getElementIn(rootElement);
+};
 
-	// Element is either template element itself or child(ren) of the template element (but not both)
-	var selection = templateElement.filter(matcher(this.elementSelector));
-	if(selection.size() === 0) {
-		selection = templateElement.select(this.elementSelector);
-	}
-	return selection;
+// Answer the data function of the receiver
+Renderer.prototype.getDataFunction = function() {
+	return this.getTemplatePath().getDataFunction();
 };
 
 // Answer whether receiver is TemplateNode
@@ -72,8 +58,8 @@ Renderer.prototype.isTemplateNode = function() {
 };
 
 // TextRenderer - Renders data as text of element
-export function TextRenderer(fieldSelector, elementSelector) {
-	Renderer.call(this, fieldSelector, elementSelector);
+export function TextRenderer(templatePath) {
+	Renderer.call(this, templatePath);
 }
 
 TextRenderer.prototype = Object.create(Renderer.prototype);
@@ -86,7 +72,7 @@ TextRenderer.prototype.render = function(templateElement, transition) {
 	}
 
 	// Render text
-	var element = this.getElement(templateElement);
+	var element = this.getElementIn(templateElement);
 	var dataFunction = this.getDataFunction();
 	if(dataFunction.isTweenFunction) {
 		if(transition) {
@@ -110,8 +96,8 @@ TextRenderer.prototype.render = function(templateElement, transition) {
 };
 
 // AttributeRenderer - Renders data as attribute of element
-export function AttributeRenderer(fieldSelector, elementSelector, attribute) {
-	Renderer.call(this, fieldSelector, elementSelector);
+export function AttributeRenderer(templatePath, attribute) {
+	Renderer.call(this, templatePath);
 	this.attribute = attribute;
 }
 
@@ -125,7 +111,7 @@ AttributeRenderer.prototype.render = function(templateElement, transition) {
 	}
 
 	// Render attribute
-	var element = this.getElement(templateElement);
+	var element = this.getElementIn(templateElement);
 	var dataFunction = this.getDataFunction();
 	if(dataFunction.isTweenFunction) {
 		if(transition) {
@@ -148,8 +134,8 @@ AttributeRenderer.prototype.render = function(templateElement, transition) {
 };
 
 // StyleRenderer - Renders data as style of element
-export function StyleRenderer(fieldSelector, elementSelector, style) {
-	Renderer.call(this, fieldSelector, elementSelector);
+export function StyleRenderer(templatePath, style) {
+	Renderer.call(this, templatePath);
 	this.style = style;
 }
 
@@ -163,7 +149,7 @@ StyleRenderer.prototype.render = function(templateElement, transition) {
 	}
 
 	// Render style
-	var element = this.getElement(templateElement);
+	var element = this.getElementIn(templateElement);
 	var dataFunction = this.getDataFunction();
 	if(dataFunction.isTweenFunction) {
 		if(transition) {
@@ -186,8 +172,8 @@ StyleRenderer.prototype.render = function(templateElement, transition) {
 };
 
 // PropertyRenderer - Renders data as property of element
-export function PropertyRenderer(fieldSelector, elementSelector, property) {
-	Renderer.call(this, fieldSelector, elementSelector);
+export function PropertyRenderer(templatePath, property) {
+	Renderer.call(this, templatePath);
 	this.property = property;
 }
 
@@ -200,7 +186,7 @@ PropertyRenderer.prototype.render = function(templateElement, transition) {
 	// be used with a property.
 
 	// Render property
-	var element = this.getElement(templateElement);
+	var element = this.getElementIn(templateElement);
 	var dataFunction = this.getDataFunction();
 	if(dataFunction.isTweenFunction) {
 		if(transition) {
@@ -224,48 +210,3 @@ PropertyRenderer.prototype.render = function(templateElement, transition) {
 		element.property(this.property, dataFunction);
 	}
 };
-
-// Answer a d3 data function for specified field selector
-export function createDataFunction(parseFieldResult) {
-	var fieldSelectors = parseFieldResult.fieldSelectors;
-	var filterReferences = parseFieldResult.filterReferences;
-
-	// Create initial value function
-	var initialValueFunction = function(d) {
-		return fieldSelectors.reduce(function(text, selector) {
-			return text !== undefined && text !== null ? text[selector] : text;
-		}, d);
-	};
-
-	// Decide if data function is tweenable (depends on last filter applied)
-	var isTweenFunction = false;
-	var lastFilterReference = filterReferences.length > 0 ? filterReferences[filterReferences.length - 1] : null;
-	if(lastFilterReference) {
-		var filter = namedRenderFilters[lastFilterReference.name];
-		if(filter && filter.isTweenFunction) {
-			isTweenFunction = true;
-		}
-	}
-
-	// Create data function based on filters and initial value
-	var dataFunction = function(d, i, nodes) {
-		var node = this;
-		return filterReferences.reduce(function(d, filterReference) {
-			var filter = namedRenderFilters[filterReference.name];
-			if(filter) {
-				var args = filterReference.args.slice(0);
-
-				// Prepend d
-				args.splice(0, 0, d);
-
-				// Append i and nodes
-				args.push(i, nodes);
-				return filter.apply(node, args);
-			}
-			return d;
-		}, initialValueFunction(d, i, nodes));
-	};
-	dataFunction.isTweenFunction = isTweenFunction;
-
-	return dataFunction;
-}
